Make build script gtk-path agnostic and update pipeline

+++ mac-setup/build-app.sh
@@ -5,13 +5,19 @@
 ## 2. call this script
 ## 3. an .app will be packed
 
+if [ $# -eq 0 ]
+then
+  echo 'Please provide the path of your gtk installation'
+  exit 1
+fi
+
 # go to script directory
 cd "${0%/*}"
 
 # delete old app, if there
 echo "clean old app"
 
-export PATH="$HOME/.new_local/bin:$HOME/gtk/inst/bin:$PATH"
+export PATH="$HOME/.local/bin:$1/inst/bin:$PATH"
 
 rm -rf ./Xournal++.app
 rm ./Xournal++.zip
@@ -40,26 +46,28 @@ export GDK_PIXBUF_MODULE_FILE="$bundle_etc/gtk-2.0/gdk-pixbuf.loaders"
 
 mkdir -p ./Xournal++.app/Contents/Resources/etc/gtk-2.0/
 gdk-pixbuf-query-loaders > ./Xournal++.app/Contents/Resources/etc/gtk-2.0/gdk-pixbuf.loaders
-sed -i -e "s:$HOME/gtk/inst/:@executable_path/../Resources/:g" ./Xournal++.app/Contents/Resources/etc/gtk-2.0/gdk-pixbuf.loaders
+sed -i -e "s:$1/inst/:@executable_path/../Resources/:g" ./Xournal++.app/Contents/Resources/etc/gtk-2.0/gdk-pixbuf.loaders
 
 echo "Copy GTK Schema"
 mkdir -p ./Xournal++.app/Contents/Resources/share/glib-2.0/schemas
-cp -rp $HOME/gtk/inst/share/glib-2.0/schemas ./Xournal++.app/Contents/Resources/share/glib-2.0/
+cp -rp $1/inst/share/glib-2.0/schemas ./Xournal++.app/Contents/Resources/share/glib-2.0/
 
 echo "Copy UI"
 cp -rp ../ui ./Xournal++.app/Contents/Resources/
 sed -i -e 's/GDK_CONTROL_MASK/GDK_META_MASK/g' ./Xournal++.app/Contents/Resources/ui/main.glade
 
-supportedLocales=("cs" "de" "it" "pl" "zh" "zh_TW" "zh_HK")
-for locale in "${supportedLocales[@]}" ; do
+for locale in */ ; do
+  if [ -f "$1/inst/share/locale/$locale/LC_MESSAGES/xournalpp.mo" ]
+  then
 	echo "Copy locale $locale"
 	mkdir -p setup/share/locale/$locale/LC_MESSAGES
 
 	# Xournal Translation
 	cp ../build/po/$locale.gmo ./Xournal++.app/Contents/Resources/share/locale/$locale/LC_MESSAGES/xournalpp.mo
 
-  # Mac Integration
-  cp $HOME/gtk/inst/share/locale/zh_CN/LC_MESSAGES/gtk-mac-integration.mo ./Xournal++.app/Contents/Resources/share/locale/$locale/LC_MESSAGES/gtk-mac-integration.mo
+    # Mac Integration
+    cp $1/inst/share/locale/zh_CN/LC_MESSAGES/gtk-mac-integration.mo ./Xournal++.app/Contents/Resources/share/locale/$locale/LC_MESSAGES/gtk-mac-integration.mo
+  fi
 done
 
 echo "Create zip"
